Move to utils and import with a fake extension
